gen: usbeth: asix: Do a fast init if link already established
The Asix driver takes the link down during init() and then brings it back up.
This commit changes this so that if a link has already been established
successfully we simply check that the link is still good.

This reduces the delay between successive network commands.

TEST=bootp; tftp ...  - see that delay is now shorter than before.

Change-Id: I044e6d12f6c175a87ee3c93bf874b497f9379c3e

BUG=chromium-os:14082
TEST=run U-boot; usb start; bootp; bootp
See that second bootp happens quickly and link stays up.

Change-Id: I79fdf4f099041ed0a077e8a833d66076c1b28402
Reviewed-on: http://gerrit.chromium.org/gerrit/200
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Tom Warren <twarren@nvidia.com>