cros: Change the shared memory structure.
This change modifies the shared memory structure such that the fields of
interest for the kernel (vbnv and nvcxt_cache) bubble up to the base of the
structure. The kernel side change requiring this modification is being
submitted under http://codereview.chromium.org/6902164/

Also, this eliminates the gpio array in the shared memory and introduces
fields to convey actual states of the bits of interest (wrote_protect,
recovery and developer mode).

The crossystem changes using this CL are being prepared under
http://codereview.chromium.org/6902172.

The kernel changes using this CL are being reviewed under
http://codereview.chromium.org/6902164/

BUG=chromium-os:12522
TEST=manual
See http://codereview.chromium.org/6902164 for test description.

Change-Id: Ifdba918ff0287e3f11851b95a0919a61370fb564