cros: Use the kernel config in the kernel image instead of the kernel script.
There are two kernel config in our kernel image. One is the kernel script wh
embedded in the bootloader address, only applied on ARM case. And the other
is embedded in the kernel config area. All the vboot related tools use the
latter. This change is to load the latter instead of former.

Previous review feedback is in http://codereview.chromium.org/6905055/.

BUG=chromium-os:14572
TEST=build u-boot-next and boot to a verified-boot image.

Change-Id: I1712e3b43bd8dc2304804f8cafa1d1b670b64a86
Reviewed-on: http://gerrit.chromium.org/gerrit/523
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>