gen: fdt: Add serial port controlled by device tree
This adds a new console serial port which is implemented by the driver
selected in the device tree. It works by redirecting to another driver.
It might be useful while we don't have proper serial device model and
the drivers that do exist don't understand the device tree.

A suitable fdt fragment might be something like:

	aliases {
		console = "/serial@70006300";
	};

	serial@70006300 {
		compatible = "nvidia,tegra20-uart", "ns16550";
		reg = <0x70006000 0x40>;
		clock-frequency = <216000000>;
	};

BUG=chromium-os:25459
TEST=build and boot U-Boot on Seaboard
Change-Id: Ia164d5ed83779ad041d04fb1fe22576c481eb62a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18257
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Gabe Black (Do Not Use) <gabeblack@google.com>