upstream: lcd: Add CONFIG_LCD_ALIGNMENT to select frame buffer alignment
The normal alignment is PAGE_SIZE, but if this is defined, we can support
other alignments.

The motivation for this change is to make the display section-aligned on
ARM so that we can easily turn off data caching for the frame buffer region
without resorting to level 2 page tables.

Series-changes: 2
- Use a more generic config CONFIG_LCD_ALIGNMENT for lcd alignment

Change-Id: Ib5bc08ee1253aa2f2ced69fec7e5c672f288b190
Reviewed-on: https://gerrit.chromium.org/gerrit/25213
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>