cros: Add assert() for debug assertions
assert() is like BUG_ON() but compiles to nothing unless DEBUG is defined.
This is useful when a condition is an error but a board reset is unlikely
to fix it, so it is better to soldier on in hope. Assertion failures should
be caught during development/test.

It turns out that assert() is defined separately in a few places in U-Boot
with various meanings.

Build errors exposed by this change (and defining DEBUG) are also fixed in
this commit.

BUG=chromium-os:11623
TEST=build U-Boot for seaboard

Change-Id: I27500491d2328405694bfd65ab991d42d9641927
Reviewed-on: http://gerrit.chromium.org/gerrit/2777
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Anton Staaf <robotboy@chromium.org>