cros: update and tidy up gbb_bmpblk
Changes include:
* CONFIG_CMD_BMP is not defined and so every reference to it is removed
* remove #ifdef guards; if any symbol is missing, build should fail

Note: debug output messes up bitmaps display, but this is intended. Most
test users do not have the privilege to access serial output, and so LCD
is their only source of debug message. We will turn off debug output on
LCD after we feel that firmware is stable enough that debug messages are
not useful to test users.

BUG=chromium-os:16508
TEST=bitmaps are displayed in developer and recovery mode on seaboard

Change-Id: I80be6544298bd6cf3db24f650583dbad247f431c
Reviewed-on: http://gerrit.chromium.org/gerrit/3067
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>