cros: Implement bootstub_entry library for entry point of bootstub firmware.
This is the entry point of verified-boot bootstub firmware. We make lots of
code into a library and minimize the function of common/cmd_vboot_bootstub
such that it is easy to adapt to the two-stop design.

Current status is it failed to boot a signed image. Need more investigation.

BUG=chromium-os:16544
TEST=build chromeos_seaboard_vboot withour error

Change-Id: Ibece02f026712a7a580290463431f7a554edd372
Reviewed-on: http://gerrit.chromium.org/gerrit/3441
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>