cros: crossystem: [1] rename to crossystem_data
Forewords:

This is the first patch of a patch series to support embedding
crossystem data into a device tree, and pass it to kernel.

The purpose is to replace a previous workaround that u-boot reserves
the last 1MB of RAM for passing crossystem data to kernel.

After the patch series is merged, you should be able to find all
crossystem data under /proc/device-tree/crossystem.

----

The kernel_shared_data is renamed to crossystem_data, which is a more
accurate description of its purpose.

BUG=chromium-os:15744
TEST=manual

1. Build kernel with device tree enabled
USE=device_tree emerge-tegra2_seaboard chromeos-kernel

2. Build u-boot-next onestop firmware

3. Check device tree content after boot
[ $(cat /proc/device-tree/crossystem/hwid) = 'ARM SEABOARD TEST 1176' ]

Change-Id: Ib532488bd01d1cd506d5f4a131f6166030619b51
Reviewed-on: http://gerrit.chromium.org/gerrit/3572
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>