config: x86: Add USB/MMC boot commands for coreboot payload
Add shortcuts to start an image on a USB key or a MMC card.
To keep the command simple, there is one limitation : if you are
booting with both MMC and USB key inserted, "usb_boot" will pick the
rootfs on the MMC instead of the USB key.

BUG=chrome-os-partner:3914
TEST=boot chromeOS from a USB key with "run usb_boot" command
boot chromeOS from a MMC card with "run mmc_boot" command

Reviewed-on: http://gerrit.chromium.org/gerrit/3679
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>

Signed-off-by: Gabe Black <gabeblack@google.com>

Conflicts:

	include/configs/coreboot.h

Change-Id: Ic8fe65245cdad3e22f776b77671244265a3a45f3
Reviewed-on: https://gerrit.chromium.org/gerrit/16895
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Tested-by: Gabe Black (Do Not Use) <gabeblack@google.com>