config: x86: Configure u-boot/coreboot to support a local and serial console simultaneously
Currently, when u-boot finds information about a framebuffer console in the
coreboot tables, it switches over to it and turns off the serial console. This
is inconvenient if you wanted to use the serial console, and very inconvenient
if the framebuffer console isn't working even though it looks like it's
configured. This change enables and configures console multiplexing so that
both the vga and serial consoles will be used if available.

BUG=chrome-os-partner:4899
TEST=Booted and verified that the serial console remained functional. Changed
the stdout environment setting to "vga" and saw the console output stop.
Changed it back to "vga,serial" and saw it start again.

Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4118
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>

Change-Id: Ida4dc6d1254d24ede7e125f4be0d7e303ffe2bd3
Reviewed-on: https://gerrit.chromium.org/gerrit/16626
Reviewed-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Tested-by: Gabe Black (Do Not Use) <gabeblack@google.com>