gen: fix enumeration on USB1 controller
When resetting the USB1 controller hub, the connection is usually bouncing
once. So just retry properly instead of giving up.

BUG=chrome-os-partner:2287
TEST=start an Aebl board with a USB stick plugged on and
observed the system is enumerating the USB key and booting on it.
Also tested on Kaen DVT2 with the USB key connected to the USB1
controller (ie the plug next to the SD card) but this requires to swap
the 2 USB controllers in the device tree to be able to use USB1.

Change-Id: I51e8dc34d3cb41a4e150c64f2eeff213546f887e
Reviewed-on: http://gerrit.chromium.org/gerrit/4095
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>