cros: fix recovery boot hangs
When the developer switch is turned off, the recovery boot hangs deep
inside call stack due to memory corruption caused by freeing a pointer
twice.

The pointer in question is allocated by VbExDiskGetInfo() and freed by
VbExDiskFreeInfo().

When VbExDiskGetInfo() finds no storage device or encounters an error, it
frees the VbDiskInfo pointer, but it still returns the pointer value to
its caller. And its caller frees the pointer again anyway.

So, when an error occurs or no storage device is available, the correct
implementation of VbExDiskGetInfo() should either not freeing and
returning the pointer, or freeing the pointer and returning NULL, but
not anything in between.

In additional to fix the free-twice bug, this patch also corrects the
return value of VbExDiskGetInfo() when there are too many storage device
available. This should be a success with warnings instead of an error.

BUG=chromium-os:17462
TEST=manually tested on Kaen

* Boot to Chrome OS on internal and external storage device through
normal, developer, and recovery boot path.
* The recovery boot path is tested with both developer switch on and
off.

Change-Id: I26026a86523b4b0f6c471c2a30a7394607a9c86b
Reviewed-on: http://gerrit.chromium.org/gerrit/4297
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>