split: gen: cros: Makes the VbExGetTimer vboot callback function board specific
The timer_get_us function which is called by the existing implementation of
VbExGetTimer isn't trivial to implement on x86. The vboot source says that the
TSC is an appropriate time source to return from this function. This change
makes the VbExGetTimer function per board so that it can use rdtsc instead of
timer_get_us.

BUG=chrome-os-partner:4552
TEST=Built and booted on x86-alex, and built on tegra2_kaen.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I335d5ae98949cd05c42dc3f30c52931bdf63a623
Reviewed-on: http://gerrit.chromium.org/gerrit/5544
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>