split: gen: cros: Refactor the CBFS implementation and add a cbfs_file_find_uncached function
If u-boot needs something out of CBFS very early before the heap is
configured, it won't be able to use the normal CBFS support which caches some
information in memory it allocates from the heap. This change adds a new
cbfs_file_find_uncached function which searchs a CBFS instance without touching
the heap. It also refactors the implementation of u-boot's CBFS support to
avoid code duplication, and allows some functions to be used without CBFS being
initialized which sets up the cache on the heap.

BUG=chrome-os-partner:5432
TEST=Built x86-alex, verified that the CBFS commands still work, and with
later changes verified that the device tree could be loaded from CBFS before
the heap is initialized.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: Ia57a5c553a0cd524073c05ca964501ccf6c5e025
Reviewed-on: http://gerrit.chromium.org/gerrit/5858
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>