cros: port test codes to current codebase
The test codes were not compiled because it depended on a prototype
codebase which is not used/compiled now.

This patch removes the dependency to the prototype codebase and let the
test codes be enabled when VBOOT_DEBUG is defined.

Because the test codes test GBB loader and we use a static buffer for
holding GBB data, this patch also moves the location of GBB buffer to
chromeos.h, which is globally accessible.

BUG=chromium-os:16542
TEST=run "vboot_test all" and see all tests pass on Kaen
TEST=run "vbexpot_test all" and see all tests pass on Kaen

Change-Id: I24c7e51cd254f31b4b28fd0becdbc77fb4ede2f8
Reviewed-on: http://gerrit.chromium.org/gerrit/5729
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>