config: x86: remove earlier prototype codes
The removed codes are belong to an earlier prototype of integrating
vboot wrapper API, which is not compiled/used anymore. Besieds, the
prototype has a known vulnerability of buffer overflow when reading
key blocks. So there is no many referential value for keeping the
prototype codes in the codebase. And keeping the prototype also adds
costs to new patches, which have to changes two implementations. All
in all, I think it is time we retire this prototype.

This patch also fixes coreboot build failure due to the removal.

BUG=chromium-os:16542
TEST=emerge-{x86-alex,tegra2_aebl} chromeos-u-boot

Reviewed-on: http://gerrit.chromium.org/gerrit/5874
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>

Signed-off-by: Gabe Black <gabeblack@google.com>

Conflicts:

	Makefile
	common/Makefile
	cros/include/hda_codec.h
	include/configs/coreboot.h

Change-Id: Ib6be9457f2dbc2daa3213c3d9c80bcdc56c520c4
Reviewed-on: https://gerrit.chromium.org/gerrit/16908
Tested-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Reviewed-by: Gabe Black (Do Not Use) <gabeblack@google.com>