lost: split: gen: cros: Assign the frame buffer address as FDT one instead of u-boot one.
Rebased-from: ed94682
Lost-file: drivers/video/tegra2.c

The framebuffer address should be specified in the device tree and not u-boot
assigned. This FDT value should be the same as the one defined in Linux kernel;
otherwise, it causes screen flicker. The FDT value overrides the framebuffer
allocated at the top of memory by board_init_f().

If the framebuffer address is not defined in the FDT, falls back to use the
address allocated by board_init_f().

BUG=chrome-os-partner:5338
TEST=build without error and boot Chrome OS on Aebl, see the screen not
corrupted and splash screen showed.

Change-Id: Ia6a996d127d74d6084900814d2c934fb95518e23
Reviewed-on: http://gerrit.chromium.org/gerrit/5877
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>