x86: Change printf to puts to avoid a buffer overflow
printf as currently implemented in u-boot has a problem where it can overflow
an internal buffer if it prints an expanded string that's too long. Our command
lines are long enough to cause this problem. A fix is coming, but in the mean
time this change replaces a problematic printf with a few calls to puts that
have the same effect. This may perform slightly better because it should avoid
a copy and scanning for format specifiers. The amount of time it actually takes
up is very tiny relative to everything else so in practice that's probably
irrelevant.

BUG=none
TEST=Used vboot_twostop on an Alex and didn't see the buffer overflow problem.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I6ccd17b892c04dea673e289c582e85d53b3930e7
Reviewed-on: http://gerrit.chromium.org/gerrit/6271
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>