lost: split: cros: x86: Pass correct VDAT address to the vboot wrapper.
Rebased-from: 4b3409c
Lost-file: arch/x86/cpu/coreboot/tables.c

On X86 systems where VDAT address is passed to the kernel through
ACPI the u-boot must retrieve the address from the coreboot table
and make it available to the vboot wrapper.

This change accomplishes just that utilizing the coreboot changes
which include the vdat area address and size into the coreboot
table.

Another modification is removing conditional compilation for
including fields in the sysinfo table. The thing is that the code
is not conditionally compiled anyways, so keeping an extra
pointer in a table even if not needed is not a problem.

BUG=chrome-os-partner:5707
TEST=manual

. program the new firmware on an Alex
. bring up ChromeOS
. try running crossystem

It was reporting errors before (when displaying data retrieved
from VDAT). It displays sensible values now.

Change-Id: Ib4090049092f7963a1b1705409d2beea8a9abfac
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6868
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6919
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>