cros: Start vboot timer at zero.
vboot API expects the firmware to provide some abstract
timestamps, which are included in the vdat area and allow the
crossystem utility to display boot timing.

This change modifies the timestamps such that the first timestamp
reading always returns zero, and the rest return incremental time
after the first reading was taken.

BUG=chrome-os-partner:5707
TEST=manual

. program the new firmware image on an Alex
. bring up ChromeOS
. run crossystem

Observe the first timestamp value set to zero and the rest of
the stamps continuously incrementing.

Change-Id: Iac4858f50dbc7576250b85630ed8e0b0cbe0b6b5
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6869
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6920
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>