x86: Fix refactoring problems introduced earlier.
It was noticed that the system fails to come up while running the
latest u-boot from the 2011.06 branch. Further investigation has
discovered that the problem is due to the fact that the
initialization of the verified boot wrapper was not handled
properly and that the IDE index reported by the IDE discovery
routine was wrong.

BUG=chromium-os:19599
TEST=manual

 . build and install the new firmware image on an Alex with a
   valid ChromeOS image
 . restart the machine
 . enter vboot_two at u-boot prompt

 observe the machine to bring up the ChromeOs login screen.

Change-Id: Icc78d682d5c81d6f08d31896c32c8cd48d0a4f2d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7030
Reviewed-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/18772
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>